+++ src/WinLzz/BWZ/MTFInverse.java
@@ -12,14 +12,14 @@ import WinLzz.BWZ.Util.SplitLinkedList;
  */
 class MTFInverse {
 
-    private short[] text;
+    private int[] text;
 
     /**
      * Creates a new instance of {@code MTFInverse}.
      *
      * @param text the text to be inverse transformed.
      */
-    MTFInverse(short[] text) {
+    MTFInverse(int[] text) {
         this.text = text;
     }
 
@@ -31,13 +31,13 @@ class MTFInverse {
      * @return the text after the inverse transform.
      */
     @Deprecated
-    short[] decode2() {
-        short[] dictionary = new short[257];
-        for (short i = 0; i < 257; i++) dictionary[i] = i;
-        short[] result = new short[text.length];
+    int[] decode2() {
+        int[] dictionary = new int[257];
+        for (int i = 0; i < 257; i++) dictionary[i] = i;
+        int[] result = new int[text.length];
         int index = 0;
-        for (short b : text) {
-            short s = dictionary[b];
+        for (int b : text) {
+            int s = dictionary[b];
             System.arraycopy(dictionary, 0, dictionary, 1, b);
             dictionary[0] = s;
             result[index++] = s;
@@ -53,11 +53,11 @@ class MTFInverse {
      *
      * @return the text after the inverse transform.
      */
-    short[] decode() {
+    int[] decode() {
         SplitLinkedList sll = new SplitLinkedList();
-        short[] result = new short[text.length];
+        int[] result = new int[text.length];
         int index = 0;
-        for (short i : text) result[index++] = sll.getAndMove(i);
+        for (int i : text) result[index++] = sll.getAndMove(i);
         return result;
     }
 }
