+++ src/WinLzz/LongHuffman/LongHuffmanDeCompressorRam.java
@@ -13,7 +13,7 @@ import java.util.HashMap;
 public class LongHuffmanDeCompressorRam {
 
     private int alphabetSize;
-    private short endSig;
+    private int endSig;
     private int maxCodeLen = 0;
     private int average = 8;
 
@@ -22,22 +22,22 @@ public class LongHuffmanDeCompressorRam {
      * shorter than {@code average} is extended by all possible combination of 0's and 1's until they reaches the
      * length {@code average}.
      */
-    private HashMap<Integer, Short> shortMap = new HashMap<>();
+    private HashMap<Integer, Integer> shortMap = new HashMap<>();
 
     /**
      * The huffman code table that records all codes that are shorter than or equal to {@code maxCodeLen}, but all
      * codes shorter than {@code maxCodeLen} is extended by all possible combination of 0's and 1's until they
      * reaches the length {@code maxCodeLen}.
      */
-    private HashMap<Integer, Short> longMap = new HashMap<>();
+    private HashMap<Integer, Integer> longMap = new HashMap<>();
 
     /**
      * The table that records all huffman symbol and their corresponding code length.
      */
-    private HashMap<Short, Integer> lengthMap = new HashMap<>();
+    private HashMap<Integer, Integer> lengthMap = new HashMap<>();
 
     private byte[] text;
-    private short[] result;
+    private int[] result;
     private int currentIndex;
     private StringBuilder builder = new StringBuilder();
 
@@ -53,15 +53,15 @@ public class LongHuffmanDeCompressorRam {
     public LongHuffmanDeCompressorRam(byte[] text, int alphabetSize, int maxLength) {
         this.text = text;
         this.alphabetSize = alphabetSize;
-        this.result = new short[maxLength];
+        this.result = new int[maxLength];
     }
 
-    private HashMap<Short, Integer> recoverLengthCode(byte[] map) {
-        HashMap<Short, Integer> lengthCode = new HashMap<>();
+    private HashMap<Integer, Integer> recoverLengthCode(byte[] map) {
+        HashMap<Integer, Integer> lengthCode = new HashMap<>();
         for (int i = 0; i < alphabetSize; i++) {
             int len = map[i] & 0xff;
             if (len != 0) {
-                lengthCode.put((short) i, len);
+                lengthCode.put(i, len);
                 if (len > maxCodeLen) maxCodeLen = len;
             }
         }
@@ -69,8 +69,8 @@ public class LongHuffmanDeCompressorRam {
         return lengthCode;
     }
 
-    private void generateIdenticalMap(HashMap<Short, String> origMap) {
-        for (short value : origMap.keySet()) {
+    private void generateIdenticalMap(HashMap<Integer, String> origMap) {
+        for (int value : origMap.keySet()) {
             String s = origMap.get(value);
             lengthMap.put(value, s.length());
 
@@ -103,7 +103,7 @@ public class LongHuffmanDeCompressorRam {
         int i = 0;
         while (true) {
             int index = Integer.parseInt(builder.substring(i, i + average), 2);
-            short value;
+            int value;
             int len;
             if (shortMap.containsKey(index)) {
                 value = shortMap.get(index);
@@ -131,7 +131,7 @@ public class LongHuffmanDeCompressorRam {
      * @param endSig The EOF character.
      * @return The uncompressed text.
      */
-    public short[] uncompress(byte[] map, short endSig) {
+    public int[] uncompress(byte[] map, int endSig) {
         this.endSig = endSig;
 
         currentIndex = 0;
@@ -140,13 +140,13 @@ public class LongHuffmanDeCompressorRam {
         longMap.clear();
         lengthMap.clear();
 
-        HashMap<Short, Integer> lengthCode = recoverLengthCode(map);
-        HashMap<Short, String> huffmanCode = LongHuffmanUtil.generateCanonicalCode(lengthCode);
+        HashMap<Integer, Integer> lengthCode = recoverLengthCode(map);
+        HashMap<Integer, String> huffmanCode = LongHuffmanUtil.generateCanonicalCode(lengthCode);
         generateIdenticalMap(huffmanCode);
         builder = Bytes.bytesToStringBuilder(text);
         uncompressToArray();
 
-        short[] rtn = new short[currentIndex];
+        int[] rtn = new int[currentIndex];
         System.arraycopy(result, 0, rtn, 0, currentIndex);
         return rtn;
     }
